Move types to separate types file
