+++ src/routes/+page.svelte
@@ -2,8 +2,9 @@
 	import MoreLink from '$lib/components/MoreLink.svelte'
 	import Splash from '$lib/components/Splash.svelte'
 	import Thumbnail from '$lib/components/Thumbnail.svelte'
-	import VideoList, { VideoListMode } from '$lib/components/VideoList.svelte'
+	import VideoList from '$lib/components/VideoList.svelte'
 	import promoInfinite from '$lib/images/promo-gb_infinite.png'
+	import { VideoListMode } from '$lib/types'
 	import type { PageData } from './$types'
 
 	export let data: PageData
